Extract RoutesEnum to paths.ts
