+++ packages/client/src/paths.ts
@@ -1,0 +1,10 @@
+export enum RoutesEnum {
+  Main = '/',
+  SignIn = '/login',
+  SignUp = '/register',
+  Profile = '/profile',
+  Game = '/game',
+  Leaderboard = '/leaderboard',
+  Forum = '/forum',
+  Error500 = '/error500',
+}
